Small documentation update and syntax, formatting cleanup in unit tests

+++ unit_test/util_gtest.cpp
@@ -21,19 +21,28 @@ namespace
 		~NonTrivialDestruct() { ; }
 	};
 
-	static_assert(oel::is_trivially_relocatable< std::pair<int *, std::unique_ptr<int>> >::value);
-	static_assert(oel::is_trivially_relocatable< std::tuple<std::unique_ptr<double>> >::value);
+	static_assert( oel::is_trivially_relocatable<
+			std::pair< int *, std::unique_ptr<double> >
+		>::value );
+	static_assert( oel::is_trivially_relocatable<
+			std::tuple< std::unique_ptr<double> >
+		>::value );
 	static_assert(oel::is_trivially_relocatable< std::tuple<> >::value);
-	static_assert( !oel::is_trivially_relocatable< std::tuple<int, NonTrivialDestruct, int> >::value );
-
-	static_assert(oel::is_trivially_relocatable< std::variant< std::unique_ptr<double>, int > >::value);
+	static_assert( !oel::is_trivially_relocatable
+		<	std::tuple<int, NonTrivialDestruct, int>
+		>::value );
+	static_assert( oel::is_trivially_relocatable<
+			std::variant< std::unique_ptr<double>, int >
+		>::value );
 	static_assert( !oel::is_trivially_relocatable< std::variant<NonTrivialDestruct> >::value );
 
 #if defined(_CPPLIB_VER) or defined(_LIBCPP_VERSION)
-	static_assert(oel::is_trivially_relocatable< std::string >::value);
+	static_assert(oel::is_trivially_relocatable<std::string>::value);
 #endif
 #if OEL_HAS_BOOST
-	static_assert(oel::is_trivially_relocatable< boost::circular_buffer< int, oel::allocator<int> > >::value);
+	static_assert( oel::is_trivially_relocatable<
+			boost::circular_buffer< int, oel::allocator<int> >
+		>::value );
 #endif
 
 	struct alignas(32) Foo { int a[24]; };
